move router up a folder. create decorators folder
