+++ src/core/decorators/render.js
