+++ src/core/decorators/send.js
